Update Combat log parser
Add support diffrent versions of game.
Update Combat log dashboards.
Update combat log Uploader desktop app.
Update Web and Desktop apps.
Add select version of game when Uplaod combat logs or see details on Web app.
Add select which combat must be upload. See which combat not support yet.

+++ src/API/CombatAnalysis.CombatParserAPI/Models/SpecializationScoreModel.cs
@@ -6,11 +6,11 @@ public class SpecializationScoreModel
 
     public double DamageScore { get; set; }
 
-    public int DamageDone { get; set; }
+    public long DamageDone { get; set; }
 
     public double HealScore { get; set; }
 
-    public int HealDone { get; set; }
+    public long HealDone { get; set; }
 
     public DateTimeOffset? Updated { get; set; }
